+++ packages/devtools/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @ai-sdk-tools/devtools
 
+## 1.0.8
+
+### Patch Changes
+
+- @ai-sdk-tools/store@1.0.8
+
+## 1.0.7
+
+### Patch Changes
+
+- @ai-sdk-tools/store@1.0.7
+
+## 1.0.6
+
+### Patch Changes
+
+- @ai-sdk-tools/store@1.0.6
+
+## 1.0.5
+
+### Patch Changes
+
+- @ai-sdk-tools/store@1.0.5
+
+## 1.0.4
+
+### Patch Changes
+
+- @ai-sdk-tools/store@1.0.4
+
+## 1.0.3
+
+### Patch Changes
+
+- @ai-sdk-tools/store@1.0.3
+
+## 1.0.2
+
+### Patch Changes
+
+- @ai-sdk-tools/store@1.0.2
+
+## 1.0.1
+
+### Patch Changes
+
+- @ai-sdk-tools/store@1.0.1
+
 ## 1.0.0
 
 ### Minor Changes